+++ cmd/aiservices/model/model_create.go
@@ -17,12 +17,12 @@ type ModelCreateCmd struct {
 
 	_ bool `cli-cmd:"create"`
 
-	Name             string      `cli-arg:"#" cli-usage:"NAME (e.g. openai/gpt-oss-120b)"`
+	Name             string      `cli-arg:"#" cli-usage:"NAME (e.g. swiss-ai/Apertus-8B-Instruct-2509)"`
 	HuggingfaceToken string      `cli-flag:"huggingface-token" cli-usage:"Huggingface token if required by the model"`
 	Zone             v3.ZoneName `cli-short:"z" cli-usage:"zone"`
 }
 
-func (c *ModelCreateCmd) CmdAliases() []string { return exocmd.GCreateAlias }
+func (c *ModelCreateCmd) CmdAliases() []string { return append(exocmd.GCreateAlias, "download") }
 func (c *ModelCreateCmd) CmdShort() string     { return "Create AI model (download from Huggingface)" }
 func (c *ModelCreateCmd) CmdLong() string {
 	return "This command creates an AI model by downloading it from Huggingface.\n\n" +