+++ cmd/aiservices/model/model_list.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/dustin/go-humanize"
 	exocmd "github.com/exoscale/cli/cmd"
 	"github.com/exoscale/cli/pkg/globalstate"
 	"github.com/exoscale/cli/pkg/output"
+	"github.com/exoscale/cli/utils"
 	v3 "github.com/exoscale/egoscale/v3"
 	"github.com/spf13/cobra"
 )
@@ -14,8 +16,9 @@ import (
 type ModelListItemOutput struct {
 	ID        v3.UUID                          `json:"id"`
 	Name      string                           `json:"name"`
+	Zone      v3.ZoneName                      `json:"zone"`
 	Status    v3.ListModelsResponseEntryStatus `json:"status"`
-	ModelSize *int64                           `json:"model_size"`
+	ModelSize string                           `json:"model_size" outputLabel:"Size"`
 }
 
 type ModelListOutput []ModelListItemOutput
@@ -41,36 +44,43 @@ Supported output template annotations: %s`,
 		strings.Join(output.TemplateAnnotations(&ModelListOutput{}), ", "))
 }
 func (c *ModelListCmd) CmdPreRun(cmd *cobra.Command, args []string) error {
-	exocmd.CmdSetZoneFlagFromDefault(cmd)
 	return exocmd.CliCommandDefaultPreRun(c, cmd, args)
 }
 func (c *ModelListCmd) CmdRun(_ *cobra.Command, _ []string) error {
 	ctx := exocmd.GContext
-	client, err := exocmd.SwitchClientZoneV3(ctx, globalstate.EgoscaleV3Client, c.Zone)
-	if err != nil {
-		return err
-	}
-	resp, err := client.ListModels(ctx)
+	client := globalstate.EgoscaleV3Client
+
+	zones, err := utils.AllZonesV3(ctx, client, c.Zone)
 	if err != nil {
 		return err
 	}
 
-	out := make(ModelListOutput, 0, len(resp.Models))
-	for _, m := range resp.Models {
-		var sizePtr *int64
-		if m.ModelSize != 0 {
-			size := m.ModelSize
-			sizePtr = &size
+	out := make(ModelListOutput, 0)
+	err = utils.ForEveryZone(zones, func(zone v3.Zone) error {
+		c := client.WithEndpoint(zone.APIEndpoint)
+		resp, err := c.ListModels(ctx)
+		if err != nil {
+			return err
 		}
-		out = append(out, ModelListItemOutput{
-			ID:        m.ID,
-			Name:      m.Name,
-			Status:    m.Status,
-			ModelSize: sizePtr,
-		})
-	}
 
-	return c.OutputFunc(&out, nil)
+		for _, m := range resp.Models {
+			var size string
+			if m.ModelSize != 0 {
+				size = humanize.IBytes(uint64(m.ModelSize))
+			}
+			out = append(out, ModelListItemOutput{
+				ID:        m.ID,
+				Name:      m.Name,
+				Zone:      zone.Name,
+				Status:    m.Status,
+				ModelSize: size,
+			})
+		}
+
+		return nil
+	})
+
+	return c.OutputFunc(&out, err)
 }
 
 func init() {